refactor(admin): migrate posts tab to canonical /social/... path
Moves the posts tab render from /admin/social/[platform]/[handle]/posts
to /social/[platform]/[handle]/posts. The admin path becomes a redirect
shim to the canonical route, matching the comments tab migration pattern.
